requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
PyYAML>=6.0.1
playwright>=1.44.0
markdownify>=0.12.1
python-dateutil>=2.9.0
aiosqlite>=0.20.0
